refactor: consolidate Shakapacker bundler/config diagnostics helpers
Follow-up cleanup from PR #3508. Doctor, SystemChecker, and
Dev::ServerManager each carried near-identical copies of the
shakapacker.yml parsing, assets-bundler detection, bundler labels, and
dev-server reload-mode helpers. Extract them into one mixin and clear the
remaining nits the #3508 review flagged.

- Add ReactOnRails::ShakapackerConfigHelpers, included by Doctor and
  SystemChecker and extended onto Dev::ServerManager (its commands live
  on `class << self`). Mirrors the existing untyped ConfigPathResolver
  mixin precedent, so no RBS signature is added.
- Unify YAML.safe_load options: SystemChecker now permits Symbol values
  (and rescues ScriptError), matching the other parsers and ServerMode.
- Memoize parsed_shakapacker_config per Doctor instance so a single
  diagnosis reads/parses the file once. ServerManager intentionally does
  not memoize: its `class << self` state would leak across specs, the
  same reason default_dev_server_mode is not memoized.
- Remove the dev_server_procfile_label alias (a pure wrapper for
  dev_server_label) and three unused ServerManager helpers
  (development_reload_feature_label, development_mode_title,
  hmr_procfile_description).
- Drop the trailing blank line from the non-HMR React Refresh help so the
  rspack config-hint bullet lines up with the bullets above; webpack
  omits the bullet instead of emitting an empty line.
- Document that the `react-refresh/babel` WEBPACK_SERVE qualifier is
  intentionally bundler-agnostic (the generated babel.config.js gates it
  on WEBPACK_SERVE for both webpack and rspack).

Net -106 lines across the diagnostics files. rbs:validate passes and the
RBS-declared ServerManager API is unchanged, so runtime type checking is
unaffected.

Closes #3538
